Raziel: Update - 14:02 - 26.7.16 Purchased Function and Checks for it.

+++ Coupon System/src/com/dao/interfaces/CouponDAO.java
@@ -14,7 +14,7 @@ public interface CouponDAO {
 	public Coupon getCoupon(long id, ClientType client) throws DaoExeption;
 	public Set<Coupon> getCoupons(long id, ClientType client) throws DaoExeption;
 	//TODO: this 2 fucntions to much generals.
-	public Set<Coupon> getCouponByPrice(long id, long price,ClientType client) throws DaoExeption;
+	public Set<Coupon> getCouponByPrice(long id, double price,ClientType client) throws DaoExeption;
 	public Set<Coupon> getCouponByType(long id, CouponType category, ClientType client) throws DaoExeption;
 	
 }
